+++ src/main/java/roomescape/service/ThemeService.java
@@ -7,7 +7,7 @@ import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import roomescape.common.exception.DuplicatedException;
 import roomescape.common.exception.ResourceInUseException;
-import roomescape.dao.ThemeDao;
+import roomescape.dao.ThemeJdbcDao;
 import roomescape.dto.request.ThemeRequestDto;
 import roomescape.dto.response.ThemeResponseDto;
 import roomescape.model.Theme;
@@ -17,9 +17,9 @@ public class ThemeService {
 
     private static final int POPULAR_DAY_RANGE = 7;
 
-    private final ThemeDao themeDao;
+    private final ThemeJdbcDao themeDao;
 
-    public ThemeService(final ThemeDao themeDao) {
+    public ThemeService(final ThemeJdbcDao themeDao) {
         this.themeDao = themeDao;
     }
